Split workspace pin helpers
