+++ Sources/cmuxApp.swift
@@ -884,11 +884,7 @@ struct cmuxApp: App {
     }
 
     private func toggleSelectedWorkspacePinned(in manager: TabManager) {
-        guard let workspace = manager.selectedWorkspace else { return }
-        _ = WorkspaceActionDispatcher.performPinAction(
-            in: manager,
-            target: .single(workspace.id)
-        )
+        WorkspacePinCommands.toggleSelectedWorkspace(in: manager)
     }
 
     private func clearSelectedWorkspaceCustomName(in manager: TabManager) {
@@ -974,18 +970,8 @@ struct cmuxApp: App {
         let workspace = manager.selectedWorkspace
         let workspaceIndex = workspace.flatMap { selectedWorkspaceIndex(in: manager, workspaceId: $0.id) }
         let windowMoveTargets = selectedWorkspaceWindowMoveTargets(in: manager)
-        let pinState = workspace.flatMap { workspace in
-            WorkspaceActionDispatcher.pinState(
-                in: manager,
-                target: .single(workspace.id)
-            )
-        }
 
-        Button(
-            (pinState?.pinned ?? true)
-                ? String(localized: "contextMenu.pinWorkspace", defaultValue: "Pin Workspace")
-                : String(localized: "contextMenu.unpinWorkspace", defaultValue: "Unpin Workspace")
-        ) {
+        Button(WorkspacePinCommands.selectedWorkspaceMenuLabel(in: manager)) {
             toggleSelectedWorkspacePinned(in: manager)
         }
         .disabled(workspace == nil)
