+++ cmuxTests/WorkspaceUnitTests.swift
@@ -2826,188 +2826,6 @@ final class WorkspaceReorderTests: XCTestCase {
     }
 }
 
-
-@MainActor
-final class WorkspaceActionDispatcherTests: XCTestCase {
-    func testSingleAndSidebarTargetsResolveTheSamePinState() throws {
-        let manager = TabManager()
-        let workspace = try XCTUnwrap(manager.tabs.first)
-
-        let singleState = try XCTUnwrap(
-            WorkspaceActionDispatcher.pinState(
-                in: manager,
-                target: .single(workspace.id)
-            )
-        )
-        let sidebarState = try XCTUnwrap(
-            WorkspaceActionDispatcher.pinState(
-                in: manager,
-                target: WorkspaceActionDispatcher.Target(
-                    workspaceIds: [workspace.id],
-                    anchorWorkspaceId: workspace.id
-                )
-            )
-        )
-
-        XCTAssertEqual(singleState, sidebarState)
-        XCTAssertTrue(singleState.pinned)
-    }
-
-    func testPinActionPinsMultipleTargetsFromAnchorState() throws {
-        let manager = TabManager()
-        let first = try XCTUnwrap(manager.tabs.first)
-        let second = manager.addWorkspace()
-        let third = manager.addWorkspace()
-        let target = WorkspaceActionDispatcher.Target(
-            workspaceIds: [second.id, third.id],
-            anchorWorkspaceId: second.id
-        )
-
-        let state = try XCTUnwrap(WorkspaceActionDispatcher.pinState(in: manager, target: target))
-        let result = WorkspaceActionDispatcher.performPinAction(state, in: manager)
-
-        XCTAssertTrue(state.pinned)
-        XCTAssertEqual(result.targetWorkspaceIds, [second.id, third.id])
-        XCTAssertEqual(result.changedWorkspaceIds, [second.id, third.id])
-        XCTAssertTrue(second.isPinned)
-        XCTAssertTrue(third.isPinned)
-        XCTAssertFalse(first.isPinned)
-        XCTAssertEqual(manager.tabs.map(\.id), [second.id, third.id, first.id])
-    }
-
-    func testCapturedPinStateKeepsLabelAndActionConsistent() throws {
-        let manager = TabManager()
-        let workspace = try XCTUnwrap(manager.tabs.first)
-        let state = try XCTUnwrap(
-            WorkspaceActionDispatcher.pinState(
-                in: manager,
-                target: .single(workspace.id)
-            )
-        )
-
-        manager.setPinned(workspace, pinned: true)
-        let result = WorkspaceActionDispatcher.performPinAction(state, in: manager)
-
-        XCTAssertTrue(state.pinned)
-        XCTAssertTrue(workspace.isPinned)
-        XCTAssertTrue(result.changedWorkspaceIds.isEmpty)
-    }
-}
-
-
-final class SidebarWorkspaceSnapshotRefreshPolicyTests: XCTestCase {
-    func testContextMenuPinChangeUpdatesDisplayedFieldsAndDefersNoisyFields() {
-        let current = Self.snapshot(
-            title: "lmao",
-            isPinned: false,
-            customColorHex: nil,
-            remoteConnectionStatusText: "Connected",
-            latestSubmittedMessage: "old message",
-            listeningPorts: [3000]
-        )
-        let next = Self.snapshot(
-            title: "lmao",
-            isPinned: true,
-            customColorHex: nil,
-            remoteConnectionStatusText: "Disconnected",
-            latestSubmittedMessage: "new message",
-            listeningPorts: [3000, 4000]
-        )
-
-        let decision = SidebarWorkspaceSnapshotRefreshPolicy.decision(
-            current: current,
-            next: next,
-            force: false,
-            contextMenuVisible: true
-        )
-
-        var expectedDisplayed = current
-        expectedDisplayed = expectedDisplayed.applyingContextMenuImmediateFields(from: next)
-        XCTAssertEqual(decision.workspaceSnapshotStorage, expectedDisplayed)
-        XCTAssertTrue(decision.workspaceSnapshotStorage?.isPinned == true)
-        XCTAssertEqual(decision.workspaceSnapshotStorage?.remoteConnectionStatusText, "Connected")
-        XCTAssertEqual(decision.workspaceSnapshotStorage?.latestSubmittedMessage, "old message")
-        XCTAssertEqual(decision.workspaceSnapshotStorage?.listeningPorts, [3000])
-        XCTAssertEqual(decision.pendingWorkspaceSnapshot, next)
-        XCTAssertTrue(decision.hasDeferredWorkspaceObservationInvalidation)
-    }
-
-    func testContextMenuImmediateOnlyChangeDoesNotCreateDeferredFlush() {
-        let current = Self.snapshot(
-            title: "old",
-            customDescription: nil,
-            isPinned: false,
-            customColorHex: nil
-        )
-        let next = Self.snapshot(
-            title: "new",
-            customDescription: "description",
-            isPinned: true,
-            customColorHex: "#C0392B"
-        )
-
-        let decision = SidebarWorkspaceSnapshotRefreshPolicy.decision(
-            current: current,
-            next: next,
-            force: false,
-            contextMenuVisible: true
-        )
-
-        XCTAssertEqual(decision.workspaceSnapshotStorage, next)
-        XCTAssertNil(decision.pendingWorkspaceSnapshot)
-        XCTAssertFalse(decision.hasDeferredWorkspaceObservationInvalidation)
-    }
-
-    func testClosedContextMenuStoresNextAndClearsPending() {
-        let current = Self.snapshot(title: "old", isPinned: false)
-        let next = Self.snapshot(title: "new", isPinned: true)
-
-        let decision = SidebarWorkspaceSnapshotRefreshPolicy.decision(
-            current: current,
-            next: next,
-            force: false,
-            contextMenuVisible: false
-        )
-
-        XCTAssertEqual(decision.workspaceSnapshotStorage, next)
-        XCTAssertNil(decision.pendingWorkspaceSnapshot)
-        XCTAssertFalse(decision.hasDeferredWorkspaceObservationInvalidation)
-    }
-
-    private static func snapshot(
-        title: String = "workspace",
-        customDescription: String? = nil,
-        isPinned: Bool = false,
-        customColorHex: String? = nil,
-        remoteConnectionStatusText: String = "Disconnected",
-        latestSubmittedMessage: String? = nil,
-        listeningPorts: [Int] = []
-    ) -> SidebarWorkspaceSnapshotBuilder.Snapshot {
-        SidebarWorkspaceSnapshotBuilder.Snapshot(
-            title: title,
-            customDescription: customDescription,
-            isPinned: isPinned,
-            customColorHex: customColorHex,
-            remoteWorkspaceSidebarText: nil,
-            remoteConnectionStatusText: remoteConnectionStatusText,
-            remoteStateHelpText: "",
-            copyableSidebarSSHError: nil,
-            latestSubmittedMessage: latestSubmittedMessage,
-            metadataEntries: [],
-            metadataBlocks: [],
-            latestLog: nil,
-            progress: nil,
-            compactGitBranchSummaryText: nil,
-            compactBranchDirectoryRow: nil,
-            branchDirectoryLines: [],
-            branchLinesContainBranch: false,
-            pullRequestRows: [],
-            listeningPorts: listeningPorts
-        )
-    }
-}
-
-
 @MainActor
 final class WorkspaceNotificationReorderTests: XCTestCase {
     func testNotificationAutoReorderDoesNotMovePinnedWorkspace() {
